Added a new symbolic linking system
This is a clever system that will tell the system that the system apps are there but are symbolically linked to the ones on the data partition. We do not use /data/app because then it will identify as two apps. One in /system/data and one in /data/app. You may wish to look over the code.